pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rtsppy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "channels>=4.0.0",
    "daphne>=4.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "markdown>=3.8",
    "django-filter>=25.1",
    "drf-spectacular>=0.27.0",
    "mtcnn-opencv>=1.0.2",
    "pillow>=11.2.1",
]
